move actions unit tests to test_charm.py
