+++ tests/unit/test_charm.py
@@ -3,15 +3,17 @@
 
 """Unit tests for the haproxy charm."""
 
+import json
 import logging
 from unittest.mock import MagicMock
 
-import ops
+import ops.testing
 import pytest
 import scenario
 
 import tls_relation
 from charm import HAProxyCharm
+from tests.unit.conftest import TEST_EXTERNAL_HOSTNAME_CONFIG
 
 logger = logging.getLogger(__name__)
 
@@ -168,3 +170,144 @@ def test_ca_certificates_removed(monkeypatch: pytest.MonkeyPatch, receive_ca_cer
 
     mock_cas_file.unlink.assert_called_once()
     assert out.app_status == ops.testing.ActiveStatus("")
+
+
+@pytest.mark.usefixtures("systemd_mock", "mocks_external_calls")
+class TestGetProxiedEndpointsAction:
+    """Test "get-proxied-endpoints" Action"""
+
+    def test_no_backend_filter(self) -> None:
+        """
+        arrange: create state with one haproxy-route relation containing
+            hostname, additional_hostnames, and paths.
+        act: trigger the get-proxied-endpoints action without a backend filter.
+        assert: returns a list of all proxied endpoints for every hostname/path combination.
+        """
+        context = ops.testing.Context(HAProxyCharm)
+
+        haproxy_route_relation = ops.testing.Relation(
+            "haproxy-route",
+            remote_app_data={
+                "hostname": f'"{TEST_EXTERNAL_HOSTNAME_CONFIG}"',
+                "additional_hostnames": json.dumps(
+                    [
+                        f"ok2.{TEST_EXTERNAL_HOSTNAME_CONFIG}",
+                        f"ok3.{TEST_EXTERNAL_HOSTNAME_CONFIG}",
+                    ]
+                ),
+                "paths": '["v1", "v2"]',
+                "ports": "[443]",
+                "protocol": '"http"',
+                "service": '"haproxy-tutorial-ingress-configurator"',
+            },
+            remote_units_data={0: {"address": '"10.75.1.129"'}},
+        )
+        charm_state = ops.testing.State(
+            relations=[haproxy_route_relation],
+            leader=True,
+            model=ops.testing.Model(name="haproxy-tutorial"),
+            app_status=ops.testing.ActiveStatus(""),
+            unit_status=ops.testing.ActiveStatus(""),
+        )
+        context.run(context.on.action("get-proxied-endpoints"), charm_state)
+
+        out = context.action_results
+
+        assert out == {
+            "endpoints": json.dumps(
+                [
+                    "https://haproxy.internal/v1",
+                    "https://haproxy.internal/v2",
+                    "https://ok2.haproxy.internal/v1",
+                    "https://ok2.haproxy.internal/v2",
+                    "https://ok3.haproxy.internal/v1",
+                    "https://ok3.haproxy.internal/v2",
+                ]
+            )
+        }
+
+    def test_no_backend_filter_no_endpoints(self) -> None:
+        """
+        arrange: create state with no haproxy-route relations.
+        act: trigger the get-proxied-endpoints action without a backend filter.
+        assert: returns an empty list.
+        """
+        context = ops.testing.Context(HAProxyCharm)
+        charm_state = ops.testing.State(
+            relations=[],
+            leader=True,
+            model=ops.testing.Model(name="haproxy-tutorial"),
+            app_status=ops.testing.ActiveStatus(""),
+            unit_status=ops.testing.ActiveStatus(""),
+        )
+        context.run(context.on.action("get-proxied-endpoints"), charm_state)
+
+        out = context.action_results
+
+        assert out == {"endpoints": "[]"}
+
+    def test_with_backend_filter(self) -> None:
+        """
+        arrange: create state with a haproxy-route relation for a specific backend.
+        act: trigger the get-proxied-endpoints action with the backend filter.
+        assert: returns a list containing the endpoint for that backend.
+        """
+        service_name = "haproxy-tutorial-ingress-configurator"
+        context = ops.testing.Context(HAProxyCharm)
+        haproxy_route_relation = ops.testing.Relation(
+            "haproxy-route",
+            remote_app_data={
+                "hostname": f'"{TEST_EXTERNAL_HOSTNAME_CONFIG}"',
+                "ports": "[443]",
+                "protocol": '"http"',
+                "service": f'"{service_name}"',
+            },
+            remote_units_data={0: {"address": '"10.75.1.129"'}},
+        )
+        charm_state = ops.testing.State(
+            relations=[haproxy_route_relation],
+            leader=True,
+            model=ops.testing.Model(name="haproxy-tutorial"),
+            app_status=ops.testing.ActiveStatus(""),
+            unit_status=ops.testing.ActiveStatus(""),
+        )
+        context.run(
+            context.on.action("get-proxied-endpoints", params={"backend": service_name}),
+            charm_state,
+        )
+
+        out = context.action_results
+
+        assert out == {"endpoints": f'["https://{TEST_EXTERNAL_HOSTNAME_CONFIG}/"]'}
+
+    def test_with_backend_filter_non_existing_backend(self) -> None:
+        """
+        arrange: create state with a haproxy-route relation for a specific backend.
+        act: trigger the get-proxied-endpoints action with a non-existing backend name.
+        assert: raises ActionFailed indicating the backend does not exist.
+        """
+        service_name = "haproxy-tutorial-ingress-configurator"
+        context = ops.testing.Context(HAProxyCharm)
+        haproxy_route_relation = ops.testing.Relation(
+            "haproxy-route",
+            remote_app_data={
+                "hostname": f'"{TEST_EXTERNAL_HOSTNAME_CONFIG}"',
+                "ports": "[443]",
+                "protocol": '"http"',
+                "service": f'"{service_name}"',
+            },
+            remote_units_data={0: {"address": '"10.75.1.129"'}},
+        )
+        charm_state = ops.testing.State(
+            relations=[haproxy_route_relation],
+            leader=True,
+            model=ops.testing.Model(name="haproxy-tutorial"),
+            app_status=ops.testing.ActiveStatus(""),
+            unit_status=ops.testing.ActiveStatus(""),
+        )
+        with pytest.raises(ops.testing.ActionFailed) as excinfo:
+            context.run(
+                context.on.action("get-proxied-endpoints", params={"backend": "random_name"}),
+                charm_state,
+            )
+        assert str(excinfo.value) == 'No backend with name "random_name"'
